default insets for intervals
